shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Shared library for TFM agents"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12",
    "fastapi>=0.135",
    "uvicorn[standard]>=0.43",
    "opentelemetry-api>=1.40",
    "opentelemetry-sdk>=1.40",
    "azure-monitor-opentelemetry>=1.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
